Move services interfaces and implementations to their own directories and namespaces
