+++ SMB3Explorer/App.xaml.cs
@@ -3,6 +3,11 @@ using System.Threading.Tasks;
 using System.Windows;
 using Microsoft.Extensions.DependencyInjection;
 using SMB3Explorer.Services;
+using SMB3Explorer.Services.ApplicationContext;
+using SMB3Explorer.Services.CsvWriterWrapper;
+using SMB3Explorer.Services.DataService;
+using SMB3Explorer.Services.NavigationService;
+using SMB3Explorer.Services.SystemInteropWrapper;
 using SMB3Explorer.ViewModels;
 using SMB3Explorer.Views;
