+++ SMB3Explorer/Services/DataService/DataServiceFranchiseCareer.cs
@@ -3,8 +3,7 @@ using Microsoft.Data.Sqlite;
 using SMB3Explorer.Models;
 using SMB3Explorer.Utils;
 
-// ReSharper disable once CheckNamespace
-namespace SMB3Explorer.Services;
+namespace SMB3Explorer.Services.DataService;
 
 public partial class DataService
 {
