+++ SMB3Explorer/Services/DataService/DataServiceFranchiseSeasons.cs
@@ -4,8 +4,7 @@ using Microsoft.Data.Sqlite;
 using SMB3Explorer.Models;
 using SMB3Explorer.Utils;
 
-// ReSharper disable once CheckNamespace
-namespace SMB3Explorer.Services;
+namespace SMB3Explorer.Services.DataService;
 
 public partial class DataService
 {
