+++ SMB3Explorer/Services/DataService/DataServiceFranchises.cs
@@ -1,13 +1,10 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
-using Microsoft.Data.Sqlite;
 using SMB3Explorer.Models;
 using SMB3Explorer.Utils;
 
-// ReSharper disable once CheckNamespace
-namespace SMB3Explorer.Services;
+namespace SMB3Explorer.Services.DataService;
 
 public partial class DataService
 {
