+++ SMB3Explorer/Services/SystemInteropWrapper/SystemInteropWrapper.cs
@@ -4,8 +4,9 @@ using System.IO;
 using System.Threading.Tasks;
 using System.Windows;
 using Microsoft.Extensions.DependencyInjection;
+using SMB3Explorer.Services.CsvWriterWrapper;
 
-namespace SMB3Explorer.Services;
+namespace SMB3Explorer.Services.SystemInteropWrapper;
 
 public class SystemInteropWrapper : ISystemInteropWrapper
 {
