+++ SMB3Explorer/ViewModels/HomeViewModel.cs
@@ -6,7 +6,10 @@ using System.Windows;
 using System.Windows.Input;
 using CommunityToolkit.Mvvm.Input;
 using SMB3Explorer.Models;
-using SMB3Explorer.Services;
+using SMB3Explorer.Services.ApplicationContext;
+using SMB3Explorer.Services.DataService;
+using SMB3Explorer.Services.NavigationService;
+using SMB3Explorer.Services.SystemInteropWrapper;
 using SMB3Explorer.Utils;
 
 namespace SMB3Explorer.ViewModels;
