+++ SMB3ExplorerTests/UtilsTests/DefaultExceptionHandlerTests.cs
@@ -1,7 +1,7 @@
 ﻿using System.Diagnostics;
 using System.Windows;
 using Moq;
-using SMB3Explorer.Services;
+using SMB3Explorer.Services.SystemInteropWrapper;
 using SMB3Explorer.Utils;
 
 namespace SMB3ExplorerTests.UtilsTests;
